Remove loginform and add the logic in the session , refactor useState
