+++ src/components/App.js
@@ -10,13 +10,14 @@ import * as helpers from '../helpers';
 
 const App = () => {
   const dispatch = useDispatch();
+  const userId = useSelector((state) => state.userId);
   const errorMsg = useSelector((state) => state.errorMsg);
   const savedUserId = localStorage.getItem('userId');
 
   const [login, setLogin] = useState('Sign In');
 
   if (savedUserId) {
-    return <Redirect to={`/home/${savedUserId}`} />;
+    return <Redirect to={`/home/${savedUserId || userId}`} />;
   }
 
   const handleClick = ({ target }) => {
