--- src/containers/LoginForm.js
@@ -1,38 +1,0 @@
-import React from 'react';
-import { useSelector } from 'react-redux';
-import PropTypes from 'prop-types';
-
-function LoginForm({
-  text, errorMsg, handleChange, sendData,
-}) {
-  const loading = useSelector((state) => state.isLoading);
-  const loadingDiv = loading === 'Loading' ? 'loader-wheel' : '';
-  return (
-    <div className="Login-forms">
-      <form className="Login-form">
-        <h3 className="Home-add-item ">{text}</h3>
-        <p className="error-msg">{errorMsg}</p>
-        <div className={loadingDiv} />
-        <input className="login-form-username" type="text" onChange={handleChange} name="username" placeholder="username" />
-        <input className="login-form-password" type="password" onChange={handleChange} name="password" placeholder="password" />
-        <button type="button" data-testid="submit-btn" onClick={sendData}>Submit</button>
-      </form>
-    </div>
-  );
-}
-
-LoginForm.defaultProps = {
-  text: '',
-  errorMsg: '',
-  handleChange: () => {},
-  sendData: () => {},
-};
-
-LoginForm.propTypes = {
-  text: PropTypes.string,
-  errorMsg: PropTypes.string,
-  handleChange: PropTypes.func,
-  sendData: PropTypes.func,
-};
-
-export default LoginForm;
